+++ src/state.cpp
@@ -59,6 +59,7 @@ void LevelSession::reset() {
     startPercent = 0;
     bestNotifiedPercent = 0;
     deathNotified = false;
+    noclipSuspected = false;
 }
 
 void EditorSession::reset() {