+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Level upload/update notifications (by [MalikHw47](https://youtube.com/@MalikHw47))
 - Per-slot custom text for upload/update notifications
+- Ignore noclip runs option
 
 ## 1.2.2